services/alp-dataflow-gen/requirements-dev.txt
setuptools-rust==1.5.2
pyarrow==11.0.0
jsonpath-ng==1.5.3
dask-sql==2023.8.0
psycopg2-binary==2.9.6
rpy2==3.5.13
pydantic >= 1.10.0, < 2.0.0
python-dotenv==1.0.0
hdbcli==2.15.19
sqlalchemy-hana==0.5.0
black==23.1.0
s3fs==2023.1.0
pycryptodome==3.17
anyio >= 3.4.0, < 4.0.0
prefect==2.14.6
prefect-dask==0.2.4
bokeh==2.4.2
prefect-shell==0.2.1
pytest==7.4.4